backend/requirements.txt
fastapi>=0.124.1,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-dotenv==1.0.1
google-adk>=0.1.0
anthropic[vertex]>=0.39.0
pydantic==2.9.2
fastmcp
mcp>=1.0.0
duckdb>=1.1.0
pandas==2.2.2
openpyxl==3.1.5
